+++ frontend/middleware/auth.ts
@@ -1,0 +1,10 @@
+import { useAuthStore } from "@/stores/AuthStore";
+
+const authStore = useAuthStore();
+
+// Reference: https://nuxt.com/docs/getting-started/routing#route-middleware
+export default defineNuxtRouteMiddleware((to, from) => {
+  if (!authStore.isAuthenticated) {
+    return navigateTo('/login/')
+  }
+});