+++ frontend/stores/AuthStore.ts
@@ -1,13 +1,16 @@
 import { defineStore } from "pinia";
+import type { User } from "@/types";
 
 interface StateShape {
   token: string | null;
+  user: User | null;
   isAuthenticated: boolean;
 }
 
-export const useAuthStore = defineStore('AuthStore', {
+export const useAuthStore = defineStore("AuthStore", {
   state: (): StateShape => ({ 
     token: null,
+    user: null,
     isAuthenticated: false, 
   }),
   getters: {
@@ -21,16 +24,27 @@ export const useAuthStore = defineStore('AuthStore', {
         this.token = null;
         this.isAuthenticated = false;
       }
+
+      const userInfo = localStorage.getItem("user");
+      if (userInfo) {
+        this.user = JSON.parse(userInfo);
+      } else {
+        localStorage.setItem("user", JSON.stringify(this.user));
+      }
     },
-    logIn(token: string) {
+    logIn(token: string, user: User) {
       this.token = token;
+      this.user = user;
       this.isAuthenticated = true;
       localStorage.setItem("token", this.token);
+      localStorage.setItem("user", JSON.stringify(this.user));
     },
     logOut() {
       this.token = "";
+      this.user = null;
       this.isAuthenticated = false;
       localStorage.setItem("token", this.token);
+      localStorage.setItem("user", JSON.stringify(this.user));
     },
   },
 })