+++ frontend/types.ts
@@ -5,6 +5,21 @@ NB: for interfaces to work correctly with useFetch, property names must match
 exactly those set in backend serializers.
 */
 
+export interface User {
+  id: number;
+  username: string;
+  telegram_id: string;
+  email: string;
+  first_name: string;
+  last_name: string;
+  profile_image: string;
+  is_active: boolean;
+  is_staff: boolean;
+  is_superuser: boolean;
+  last_login: Date;
+  date_joined: Date;
+}
+
 export interface Folder {
   id: ID;
   user: ID;